Add new feature to CLI

+++ milvus_cli/scripts/milvus_cli.py
@@ -7,7 +7,7 @@ parentdir = os.path.dirname(currentdir)
 sys.path.append(parentdir)
 from utils import PyOrm, Completer, getPackageVersion
 from Fs import readCsvFile
-from Validation import validateParamsByCustomFunc, validateCollectionParameter, validateIndexParameter, validateSearchParams, validateQueryParams
+from Validation import validateParamsByCustomFunc, validateCollectionParameter, validateIndexParameter, validateSearchParams, validateQueryParams, validateQueryParams
 from Types import ParameterException, ConnectException
 from Types import MetricTypes, IndexTypesMap, IndexTypes
 
@@ -121,33 +121,53 @@ def indexProgress(obj, collection, index):
 
 @cli.command()
 @click.option('-c', '--collection', 'collection', help='The name of collection to load.')
+@click.option('-p', '--partition', 'partition', help='[Optional, Multiple] - The name of partition to load.', default=[], multiple=True)
 @click.pass_obj
-def load(obj, collection):
-    """Load specified collection."""
+def load(obj, collection, partition):
+    """Load specified collection and partitions."""
     try:
         validateParamsByCustomFunc(
             obj.getTargetCollection, 'Collection Name Error!', collection)
-        result = obj.loadCollection(collection)
+        for partitionName in partition:
+            validateParamsByCustomFunc(
+                obj.getTargetPartition, 'Partition Name Error!', collection, partitionName)
+        if partition:
+            result = obj.loadPartitions(collection, partition)
+        else:
+            result = obj.loadCollection(collection)
     except Exception as e:
         click.echo(message=e, err=True)
     else:
-        click.echo("""Load Collection '{}' successfully""".format(collection))
+        if partition:
+            click.echo(f"""Load {collection}'s partitions {partition} successfully""")
+        else:
+            click.echo(f"""Load Collection {collection} successfully""")
         click.echo(result)
 
 
 @cli.command()
 @click.option('-c', '--collection', 'collection', help='The name of collection to be released.')
+@click.option('-p', '--partition', 'partition', help='[Optional, Multiple] - The name of partition to released.', default=[], multiple=True)
 @click.pass_obj
-def release(obj, collection):
-    """Release specified collection."""
+def release(obj, collection, partition):
+    """Release specified collection and partitions."""
     try:
         validateParamsByCustomFunc(
             obj.getTargetCollection, 'Collection Name Error!', collection)
-        result = obj.releaseCollection(collection)
+        for partitionName in partition:
+            validateParamsByCustomFunc(
+                obj.getTargetPartition, 'Partition Name Error!', collection, partitionName)
+        if partition:
+            result = obj.releasePartitions(collection, partition)
+        else:
+            result = obj.releaseCollection(collection)
     except Exception as e:
         click.echo(message=e, err=True)
     else:
-        click.echo("""Release Collection '{}' successfully""".format(collection))
+        if partition:
+            click.echo(f"""Release {collection}'s partitions {partition} successfully""")
+        else:
+            click.echo(f"""Release Collection {collection} successfully""")
         click.echo(result)
 
 
@@ -159,7 +179,7 @@ def listDetails(obj):
 
 
 @listDetails.command()
-@click.option('--timeout', '-t', 'timeout', help="[Optional] - An optional duration of time in seconds to allow for the RPC. When timeout is set to None, client waits until server response or error occur.", default=None)
+@click.option('--timeout', '-t', 'timeout', help="[Optional] - An optional duration of time in seconds to allow for the RPC. When timeout is not set, client waits until server response or error occur.", default=None, type=float)
 @click.option('--show-loaded', '-l', 'showLoaded', help="[Optional] - Only show loaded collections.", default=False)
 @click.pass_obj
 def collections(obj, timeout, showLoaded):
@@ -202,7 +222,7 @@ def indexes(obj, collection):
 @cli.group('describe', no_args_is_help=False)
 @click.pass_obj
 def describeDetails(obj):
-    """Describe collection or partition."""
+    """Describe collection, partition and index."""
     pass
 
 
@@ -243,6 +263,26 @@ def describePartition(obj, collectionName, partition):
         click.echo(f"Error when getting collection by name!\n{str(e)}")
     else:
         click.echo(obj.getPartitionDetails(collection, partition))
+
+
+@describeDetails.command('index')
+@click.option('-c', '--collection', 'collectionName', help='The name of collection.')
+@click.pass_obj
+def describeIndex(obj, collectionName):
+    """
+    Describe index.
+
+    Example:
+
+        milvus_cli > describe index -c car
+    """
+    try:
+        obj.checkConnection()
+        collection = obj.getTargetCollection(collectionName)
+    except Exception as e:
+        click.echo(f"Error when getting collection by name!\n{str(e)}")
+    else:
+        click.echo(obj.getIndexDetails(collection))
 
 
 @cli.group('create', no_args_is_help=False)
@@ -365,7 +405,7 @@ def deleteObject(obj):
 
 @deleteObject.command('collection')
 @click.option('-c', '--collection', 'collectionName', help='The name of collection to be deleted.')
-@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is set to None, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
+@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is not set, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
 @click.pass_obj
 def deleteCollection(obj, collectionName, timeout):
     """
@@ -393,7 +433,7 @@ def deleteCollection(obj, collectionName, timeout):
 @deleteObject.command('partition')
 @click.option('-c', '--collection', 'collectionName', help='Collection name')
 @click.option('-p', '--partition', 'partition', help='The name of partition.')
-@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is set to None, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
+@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is not set, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
 @click.pass_obj
 def deletePartition(obj, collectionName, partition, timeout):
     """
@@ -420,7 +460,7 @@ def deletePartition(obj, collectionName, partition, timeout):
 
 @deleteObject.command('index')
 @click.option('-c', '--collection', 'collectionName', help='Collection name')
-@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is set to None, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
+@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is not set, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
 @click.pass_obj
 def deleteIndex(obj, collectionName, timeout):
     """
@@ -441,6 +481,43 @@ def deleteIndex(obj, collectionName, timeout):
         click.echo(f"Error occurred when get collection by name!\n{str(e)}")
     else:
         result = obj.dropIndex(collectionName, timeout)
+        click.echo("Drop index successfully!") if not result else click.echo(
+            "Drop index failed!")
+
+
+@deleteObject.command('entities')
+@click.option('-c', '--collection', 'collectionName', help='Collection name.')
+@click.option('-p', '--partition', 'partitionName', help='[Optional] - Name of partitions that contain entities.', default='')
+@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is not set, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
+@click.pass_obj
+def deleteEntities(obj, collectionName, partitionName, timeout):
+    """
+    Delete entities with an expression condition. And return results to show which primary key is deleted successfully.
+
+    Example:
+
+        milvus_cli > delete entities -c car
+
+        The expression to specify entities to be deleted, such as "film_id in [ 0, 1 ]": film_id in [ 0, 1 ]
+
+        You are trying to delete the entities of collection. This action cannot be undone!
+
+        Do you want to continue? [y/N]: y
+    """
+    expr = click.prompt('''The expression to specify entities to be deleted, such as "film_id in [ 0, 1 ]"''')
+    click.echo(
+        "You are trying to delete the entities of collection. This action cannot be undone!\n")
+    if not click.confirm('Do you want to continue?'):
+        return
+    try:
+        obj.checkConnection()
+        obj.getTargetCollection(collectionName)
+    except Exception as e:
+        click.echo(f"Error occurred when get collection by name!\n{str(e)}")
+    else:
+        partitionValue = partitionName if partitionName else None
+        timeoutValue = timeout if timeout else None
+        result = obj.deleteEntities(expr, collectionName, partitionValue, timeoutValue)
         click.echo("Drop index successfully!") if not result else click.echo(
             "Drop index failed!")
 
@@ -628,7 +705,7 @@ def query(obj):
 @cli.command('import')
 @click.option('-c', '--collection', 'collectionName', help='The name of collection to be imported.')
 @click.option('-p', '--partition', 'partitionName', help='[Optional] - The partition name which the data will be inserted to, if partition name is not passed, then the data will be inserted to “_default” partition.', default=None)
-@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is set to None, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
+@click.option('-t', '--timeout', 'timeout', help='[Optional] - An optional duration of time in seconds to allow for the RPC. If timeout is not set, the client keeps waiting until the server responds or an error occurs.', default=None, type=float)
 @click.argument('path')
 @click.pass_obj
 def importData(obj, collectionName, partitionName, timeout, path):
@@ -659,6 +736,64 @@ def importData(obj, collectionName, partitionName, timeout, path):
         click.echo("Error!\n{}".format(str(e)))
     else:
         click.echo(f"Insert successfully.\nTotal entities: {entitiesNum}")
+
+
+@cli.command('calc')
+@click.pass_obj
+def calcDistance(obj):
+    """Calculate distance between two vector arrays."""
+    leftVectorMeta = {}
+    if click.confirm('Import left operator vectors from existing collection?'):
+        left_ids = click.prompt("The vectors' ids on the left of operator")
+        left_collection = click.prompt("The vectors' collection name on the left of operator")
+        left_partition = click.prompt("The vectors' partition name on the left of operator")
+        left_field = click.prompt("The vectors' field name on the left of operator")
+        leftVectorMeta['vec_type'] = 'import'
+        leftVectorMeta['ids'] = left_ids
+        leftVectorMeta['collection'] = left_collection
+        leftVectorMeta['partition'] = left_partition
+        leftVectorMeta['field'] = left_field
+    else:
+        left_type = click.prompt("The vector's type", type=click.Choice(['float_vectors', 'bin_vectors']))
+        left_vectors = click.prompt("Left vectors")
+        leftVectorMeta['vec_type'] = 'raw'
+        leftVectorMeta['type'] = left_type
+        leftVectorMeta['vectors'] = left_vectors
+    rightVectorMeta = {}
+    if click.confirm('Import right operator vectors from existing collection?'):
+        right_ids = click.prompt("The vectors' ids on the right of operator")
+        right_collection = click.prompt("The vectors' collection name on the right of operator")
+        right_partition = click.prompt("The vectors' partition name on the right of operator")
+        right_field = click.prompt("The vectors' field name on the right of operator")
+        rightVectorMeta['vec_type'] = 'import'
+        rightVectorMeta['ids'] = right_ids
+        rightVectorMeta['collection'] = right_collection
+        rightVectorMeta['partition'] = right_partition
+        rightVectorMeta['field'] = right_field
+    else:
+        right_type = click.prompt("The vector's type", type=click.Choice(['float_vectors', 'bin_vectors']))
+        right_vectors = click.prompt("Right vectors")
+        rightVectorMeta['vec_type'] = 'raw'
+        rightVectorMeta['type'] = right_type
+        rightVectorMeta['vectors'] = right_vectors
+    metric_type = click.prompt('Supported metric type. Default is "L2"', default="L2", type=click.Choice(MetricTypes))
+    sqrt = None
+    if metric_type in ["L2"]:
+        sqrt = click.prompt('sqrt', type=bool, default=False)
+    dim = None
+    if metric_type in ["HAMMING", "TANIMOTO"]:
+        dim = click.prompt('Set this value if dimension is not a multiple of 8, otherwise the dimension will be calculted by list length', type=int, default=None)
+    timeout = click.prompt('Timeout(optional)', default='')
+    try:
+        calcParams = validateQueryParams(
+            leftVectorMeta, rightVectorMeta, metric_type, sqrt, dim, timeout)
+        result = obj.calcDistance(calcParams['vectors_left'], calcParams['vectors_right'], calcParams['params'], calcParams['timeout'])
+    except ParameterException as pe:
+        click.echo("Error!\n{}".format(str(pe)))
+    except ConnectException as ce:
+        click.echo("Error!\n{}".format(str(ce)))
+    else:
+        click.echo(result)
 
 
 @cli.command('exit')
